requirements.txt
tensorflow
pandas
numpy
matplotlib
seaborn
scikit-learn
fastapi
uvicorn
python-multipart
pillow